requirements.txt
# AgreeCulture — Python Dependencies
# Install with: pip install -r requirements.txt

flask==3.0.3
werkzeug==3.0.3
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
Pillow==10.3.0

# AI Providers (both free)
# 1. Google Gemini - https://aistudio.google.com/
google-genai>=1.0.0

# 2. OpenRouter - https://openrouter.ai/ (Llama, Mistral, Gemma — free models)
# Uses standard HTTP requests — no extra package needed (requests is included with Flask)

# .env file support
python-dotenv>=1.0.0
